chore: Notifications refactoring to separate components (#2874)
